streamlit/requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
narwhals==1.13.5
numpy==1.26.4
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.3
Pillow==9.5.0
Plost==0.2.5
plotly==5.16.0
protobuf==4.25.5
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
streamlit==1.25.0
streamlit-nightly==1.40.2.dev20241113
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
zipp==3.21.0
